Move Footer to legacy
