+++ react/Footer.js
@@ -3,11 +3,11 @@ import React, { Component } from 'react'
 import classNames from 'classnames'
 import { Container } from 'vtex.store-components'
 
-import FooterBadgeList from './components/FooterBadgeList'
-import FooterLinksMatrix from './components/FooterLinksMatrix'
-import FooterPaymentFormMatrix from './components/FooterPaymentFormMatrix'
-import FooterVtexLogo from './components/FooterVtexLogo'
-import FooterSocialNetworkList from './components/FooterSocialNetworkList'
+import FooterBadgeList from './components/legacy/FooterBadgeList'
+import FooterLinksMatrix from './components/legacy/FooterLinksMatrix'
+import FooterPaymentFormMatrix from './components/legacy/FooterPaymentFormMatrix'
+import FooterVtexLogo from './components/legacy/FooterVtexLogo'
+import FooterSocialNetworkList from './components/legacy/FooterSocialNetworkList'
 import { objectLikeBadgeArray, objectLikeLinkArray } from './propTypes'
 
 import footer from './footer.css'
