+++ react/components/legacy/FooterBadgeList.js
