+++ react/components/legacy/FooterSocialNetworkList.js
@@ -3,7 +3,7 @@ import React from 'react'
 
 import footerList from './footerList'
 import withImage from './withImage'
-import footer from '../footer.css'
+import footer from '../../footer.css'
 
 /**
  * Shows an image for an specific social network
