+++ react/components/legacy/withImage.js
