+++ store/interfaces.json
@@ -1,5 +1,5 @@
 {
   "footer": {
-    "component": "index"
+    "component": "Footer"
   }
 }
